requirements.txt
streamlit
pandas
openai
python-docx
openpyxl
PyPDF2
reportlab

